move product fetch function to context
